users/pyproject.toml
[project]
name = "python-intermediate"
version = "0.1.0"
description = ""
authors = [
    {name = "Sar0102",email = "sarvarchikkamilov2@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (==0.115.12)",
    "uvicorn (==0.34.2)",
    "bcrypt (==4.3.0)",
    "python-jose (==3.5.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "strawberry-graphql[debug-server] (>=0.270.4,<0.271.0)",
    "secretstorage (>=3.3.3,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "alembic (>=1.16.1,<2.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "pytest (>=8.4.0,<9.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "sentry-sdk[fastapi] (>=2.29.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
